+++ DepRegAttributes.Analyzer/ServiceProviderAttributeAnalyzer.cs
@@ -34,8 +34,20 @@ namespace DepRegAttributes.Analyzer
                 DiagnosticSeverity.Warning,
                 isEnabledByDefault: true);
 
+        private static readonly DiagnosticDescriptor PotentialBadParameter =
+            new($"ServiceRegistration004",
+                "Potential bad parameter",
+                "{0}",
+                "ServiceRegistration",
+                DiagnosticSeverity.Warning,
+                isEnabledByDefault: true);
+
         public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics =>
-            ImmutableArray.Create(ServiceNotOnImplementation, InvalidImplementationType, InvalidAutoIncludedService);
+            ImmutableArray.Create(
+                ServiceNotOnImplementation, 
+                InvalidImplementationType, 
+                InvalidAutoIncludedService,
+                PotentialBadParameter);
 
         public override void Initialize(AnalysisContext context)
         {
@@ -75,7 +87,7 @@ namespace DepRegAttributes.Analyzer
                                 location,
                                 message));
                     }
-                    else if (service.DeclaredAccessibility != Accessibility.Public)
+                    else if (service.DeclaredAccessibility is not (Accessibility.Public or Accessibility.Internal))
                     {
                         if (service.Name == $"I{implementation.Name}" && location == attribute.GetLocation())
                         {
@@ -83,7 +95,7 @@ namespace DepRegAttributes.Analyzer
                                 Diagnostic.Create(
                                     InvalidAutoIncludedService,
                                     location,
-                                    $"{service.Name} is not public, so it will not be automatically included for {implementation.Name}"));
+                                    $"{service.Name} is not public or internal, so it will not be automatically included for {implementation.Name}"));
                         }
                         else
                         {
@@ -91,43 +103,83 @@ namespace DepRegAttributes.Analyzer
                                 Diagnostic.Create(
                                     ServiceNotOnImplementation,
                                     location,
-                                    $"{service.Name} is private, you cannot use it as a service type"));
+                                    $"{service.Name} is not public or internal, you cannot use it as a service type"));
                         }
                     }
                 }
 
-                if (implementation.IsAbstract)
+
+                AnalyzeImplementation(context, implementation, attribute);
+                AnalyzeProperties(context, attribute);
+            }
+        }
+
+        private void AnalyzeProperties(SyntaxNodeAnalysisContext context, AttributeSyntax attribute)
+        {
+            if (attribute.ArgumentList is null)
+                return;
+
+            foreach (var argument in attribute.ArgumentList.Arguments)
+            {
+                if (argument.NameEquals is null || argument.NameEquals.Name.Identifier.Text is not ("Tag" or "Key"))
+                    continue;
+
+                if(argument.Expression is ImplicitArrayCreationExpressionSyntax)
                 {
                     context.ReportDiagnostic(
-                        Diagnostic.Create(
-                            InvalidImplementationType,
-                            implementation.Locations[0],
-                            $"An abstract class cannot use the {attribute.Name} attribute"));
+                    Diagnostic.Create(
+                        PotentialBadParameter,
+                        argument.GetLocation(),
+                        $"Using an array as a {argument.NameEquals.Name.Identifier.Text} will result in a reference comparison, you will not be able to resolve it."));
                 }
-                else if (implementation.IsStatic)
-                {
-                    context.ReportDiagnostic(
-                        Diagnostic.Create(
-                            InvalidImplementationType,
-                            implementation.Locations[0],
-                            $"A static class cannot use the {attribute.Name} attribute"));
-                }
-                else if (!implementation.Constructors.Any(c => c.DeclaredAccessibility == Accessibility.Public))
-                {
-                    context.ReportDiagnostic(
-                        Diagnostic.Create(
-                            InvalidImplementationType,
-                            implementation.Locations[0],
-                            $"You need at least one public constructor to use the {attribute.Name} attribute"));
-                }
-                else if (implementation.DeclaredAccessibility != Accessibility.Public)
-                {
-                    context.ReportDiagnostic(
-                        Diagnostic.Create(
-                            InvalidImplementationType,
-                            implementation.Locations[0],
-                            $"Only public classes can use the {attribute.Name} attribute"));
-                }
+            }
+        }
+
+        private void AnalyzeImplementation(
+            SyntaxNodeAnalysisContext context, 
+            INamedTypeSymbol implementation, 
+            AttributeSyntax attribute)
+        {
+            if (implementation.IsUnboundGenericType)
+            {
+                context.ReportDiagnostic(
+                    Diagnostic.Create(
+                        InvalidImplementationType,
+                        implementation.Locations[0],
+                        $"An unbound generic cannot use the {attribute.Name} attribute"));
+            }
+
+            if (implementation.IsAbstract)
+            {
+                context.ReportDiagnostic(
+                    Diagnostic.Create(
+                        InvalidImplementationType,
+                        implementation.Locations[0],
+                        $"An abstract class cannot use the {attribute.Name} attribute"));
+            }
+            else if (implementation.IsStatic)
+            {
+                context.ReportDiagnostic(
+                    Diagnostic.Create(
+                        InvalidImplementationType,
+                        implementation.Locations[0],
+                        $"A static class cannot use the {attribute.Name} attribute"));
+            }
+            else if (!implementation.Constructors.Any(c => c.DeclaredAccessibility is Accessibility.Public or Accessibility.Internal))
+            {
+                context.ReportDiagnostic(
+                    Diagnostic.Create(
+                        InvalidImplementationType,
+                        implementation.Locations[0],
+                        $"You need at least one public constructor to use the {attribute.Name} attribute"));
+            }
+            else if (implementation.DeclaredAccessibility is not (Accessibility.Public or Accessibility.Internal))
+            {
+                context.ReportDiagnostic(
+                    Diagnostic.Create(
+                        InvalidImplementationType,
+                        implementation.Locations[0],
+                        $"Only public classes can use the {attribute.Name} attribute"));
             }
         }
     }